# Conflicts:
#	core/application/src/main/java/org/naho/grammar/result/GrammarResult.java
#	core/application/src/main/java/org/naho/question/command/CreateSpeakingQuestionCommand.java
#	core/application/src/main/java/org/naho/question/usecase/CreateSpeakingQuestionUseCase.java
#	core/application/src/main/java/org/naho/vocabulary/result/VocabularyResult.java
#	core/bootstrap/src/main/java/org/naho/config/application/GrammarConfig.java
#	core/bootstrap/src/main/java/org/naho/config/application/SpeakingQuestionConfig.java
#	core/domain/src/main/java/org/naho/question/model/SpeakingQuestion.java
#	core/infrastructure/src/main/java/org/naho/vocabulary/adapter/VocabularyRepositoryAdapter.java
#	core/presentation/src/main/java/org/naho/question/controller/v1/VocabularyQuestionController.java
#	core/presentation/src/main/java/org/naho/vocabulary/dto/response/VocabularyResponse.java

+++ core/application/src/main/java/org/naho/vocabulary/port/in/ExportVocabularyInputPort.java
@@ -3,7 +3,7 @@ package org.naho.vocabulary.port.in;
 import java.io.ByteArrayInputStream;
 
 public interface ExportVocabularyInputPort {
-    ByteArrayInputStream exportByQuestion(int questionId);
+    ByteArrayInputStream exportByQuestion(Long questionId);
 
-    ByteArrayInputStream exportByObjective(int objectiveId);
+    ByteArrayInputStream exportByObjective(Long objectiveId);
 }
