+++ core/application/src/main/java/org/naho/vocabulary/port/in/GetVocabulariesOfObjectiveInputPort.java
@@ -3,5 +3,5 @@ package org.naho.vocabulary.port.in;
 import org.naho.vocabulary.result.VocabulariesOfObjectiveResult;
 
 public interface GetVocabulariesOfObjectiveInputPort {
-    VocabulariesOfObjectiveResult getVocabularyListOfObjective(int objectiveId);
+    VocabulariesOfObjectiveResult getVocabularyListOfObjective(Long objectiveId);
 }
